# Grampstype set