--- a/gramps/gen/lib/grampstype.py
+++ b/gramps/gen/lib/grampstype.py
@@ -151,7 +151,7 @@ class GrampsType(metaclass=GrampsTypeMeta):
         else:
             self.__string = ""
 
-    def __set_dict(self, value):
+    def set_dict(self, value):
         "Set the value/string properties from a dict."
         val, strg = self._DEFAULT, ""
         if value:
@@ -161,7 +161,7 @@ class GrampsType(metaclass=GrampsTypeMeta):
         self.__value = val
         self.__string = strg
 
-    def __set_tuple(self, value):
+    def set_tuple(self, value):
         "Set the value/string properties from a tuple."
         val, strg = self._DEFAULT, ""
         if value:
@@ -171,12 +171,12 @@ class GrampsType(metaclass=GrampsTypeMeta):
         self.__value = val
         self.__string = strg
 
-    def __set_int(self, value):
+    def set_int(self, value):
         "Set the value/string properties from an integer."
         self.__value = value
         self.__string = ""
 
-    def __set_instance(self, value):
+    def set_instance(self, value):
         "Set the value/string properties from another grampstype."
         self.__value = value.value
         if self.__value == self._CUSTOM:
@@ -184,7 +184,7 @@ class GrampsType(metaclass=GrampsTypeMeta):
         else:
             self.__string = ""
 
-    def __set_str(self, value):
+    def set_str(self, value):
         "Set the value/string properties from a string."
         self.__value = self._S2IMAP.get(value, self._CUSTOM)
         if self.__value == self._CUSTOM:
@@ -195,15 +195,15 @@ class GrampsType(metaclass=GrampsTypeMeta):
     def set(self, value):
         "Set the value/string properties from the passed in value."
         if isinstance(value, dict):
-            self.__set_dict(value)
+            self.set_dict(value)
         elif isinstance(value, tuple):
-            self.__set_tuple(value)
+            self.set_tuple(value)
         elif isinstance(value, int):
-            self.__set_int(value)
+            self.set_int(value)
         elif isinstance(value, self.__class__):
-            self.__set_instance(value)
+            self.set_instance(value)
         elif isinstance(value, str):
-            self.__set_str(value)
+            self.set_str(value)
         else:
             self.__value = self._DEFAULT
             self.__string = ""
@@ -343,5 +343,5 @@ class GrampsType(metaclass=GrampsTypeMeta):
     def __ne__(self, value):
         return not self.__eq__(value)
 
-    value = property(__int__, __set_int, None, "Returns or sets integer value")
-    string = property(__str__, __set_str, None, "Returns or sets string value")
+    value = property(__int__, set_int, None, "Returns or sets integer value")
+    string = property(__str__, set_str, None, "Returns or sets string value")