--- a/gramps/gui/views/treemodels/notemodel.py
+++ b/gramps/gui/views/treemodels/notemodel.py
@@ -118,7 +118,7 @@ class NoteModel(FlatBaseModel):
 
     def column_type(self, data):
         """Return the type of the Note in readable format."""
-        return NoteType.get_str(data["type"])
+        return NoteType.get_str(data.type)
 
     def column_preview(self, data):
         """Return a shortend version of the Note's text."""