--- a/gramps/gui/views/treemodels/placemodel.py
+++ b/gramps/gui/views/treemodels/placemodel.py
@@ -188,8 +188,7 @@ class PlaceBaseModel:
         return data["gramps_id"]
 
     def column_type(self, data):
-        pt = from_dict(data["place_type"])
-        return str(pt)
+        return PlaceType.get_str(data["place_type"])
 
     def column_code(self, data):
         return data["code"]