Make the `GrampsType.__set` methods public

Based on a comment by @dsblank in [PR #1825](https://github.com/gramps-project/gramps/pull/1825#discussion_r1895209505) 
This PR builds on PR #1825

No attempt has been made to convert calling code. That could be tackled if there is agreement on the approach. 

CC @Nick-Hall for comment on the approach