@dsblank 2024-12-27
@Nick-Hall, @stevenyoungs has made a couple of PRs to choose from: this one, or #1825. I think this one is better, but not sure if there is a difference in functionality of a private static class method.