@stevenyoungs 2024-12-28
> @Nick-Hall, @stevenyoungs has made a couple of PRs to choose from: this one, or #1825. I think this one is better, but not sure if there is a difference in functionality of a private static class method.

#1825 contains a bug fix for #1726. 
This PR, #1830, adds one additional [commit](https://github.com/gramps-project/gramps/pull/1830/commits/5c44298030c5f992114313b6f30c5e744c731ddb) on top of #1825, This additional commit makes the `GrampsType.__set_*` methods public.