@Nick-Hall 2025-01-09
I don't see the point of the PR.  Why do you want to make these methods public?