@Nick-Hall 2025-01-11
The original code uses a standard technique for overloading methods in python.  Four methods (now five) were combined into a single `set` method to make the public interface simpler.  This does come with a slight loss in performance, but that should be negligible.

Note:  The python typing module includes the [overload](https://docs.python.org/3/library/typing.html#typing.overload) decorator for such methods.