@QuLogic 2025-01-12
> Note: The python typing module includes the [overload](https://docs.python.org/3/library/typing.html#typing.overload) decorator for such methods.

This is purely for describing the types of the function; it doesn't do anything at runtime. For runtime dispatch, that's [`functools.singledispatch`](https://docs.python.org/3/library/functools.html#functools.singledispatch).