@stevenyoungs 2025-01-12
@QuLogic thanks for the information on `singledispatch`. I've refactored my PR to use this technique. I would appreciate if you can check to see if my code can be improved. I did not particularly like having the `def set_instance(self, value: GrampsType)` function outside the class but I could not find any other way. I then had to have a `set_instance` method in the class in order to set the private properties. Introducing a new super class seemed overkill.

@Nick-Hall I don't think `singledispatch` \ `singledispatchmethod` are currently used in the gramps codebase. It allows us to remove the
```
is isinstance()
elif is instance()
elif is instance()
```
construct. This seems like a positive change. Any comments?