@Nick-Hall 2025-01-13
@stevenyoungs @QuLogic I've squashed the commits, refactored the `__set_instance` method back into the class, and changed how the method is registered.  If you have no further suggestions then I think that this is ready to merge.