@stevenyoungs 2025-01-13
I like your solution to moving `__set_instance` back into the class.
Good to merge from my point of view.