@prculley 2025-02-12
I am sorry for the late comment; been busy with other things.
I just encountered this change when I tried to debug some of my addon code; I was initializing a PlaceType with the DataDict.
I was shocked to see that it ran through what seemed like hundreds of lines of code in functools, as opposed to the original if/then/else style from GrampsType.  This seems like it might be a considerable slowdown for code that sets/writes new data, like imports (which I have already observed to be very slow).

Has anyone done a performance comparison?