@dsblank 2025-02-12
I don't know if we did a specific performance test on `singledispatch`.  It sounds like a good idea since we are considering the decorator for other uses in the codebase.