@prculley 2025-02-20
I just got around to making some tests on performance.  I made a test file that instantiated all of our GrampsTypes with each of the potential 'set' values (dict, tuple, int, string, and type).  I ran this test 100000 times.
with this PR (normal Gramps60 branch), it took 59.5 seconds.
reverting this PR, it took 15.1 seconds.
So about 4 times linger with this PR.

For a more real life test, I thought that doing some imports would involve GrampsType instantiations.
I enhanced the import_test by adding example.gramps for a larger file and then ran the tests.
with PR, it took 26.3 seconds
without, it took 25.8 seconds
or about a 2% slowdown.
Just for fun I also ran import_test this way on gramps52 branch, about 22.8 seconds, which is about 15% faster than gramps60.